+++ src/BizHawk.Emulation.Cores/Computers/SinclairSpectrum/ZXSpectrum.Messaging.cs
@@ -391,7 +391,7 @@ namespace BizHawk.Emulation.Cores.Computers.SinclairSpectrum
 			int end = _machine.TapeDevice.DataBlocks[_machine.TapeDevice.CurrentDataBlockIndex].DataPeriods.Count;
 			double p = 0;
 			if (end != 0)
-				p = ((double)pos / (double)end) * 100.0;
+				p = (pos / (double)end) * 100.0;
 
 			sb.Append(p.ToString("N0", NumberFormatInfo.InvariantInfo) + "%");
 			SendMessage(sb.ToString().TrimEnd('\n'), MessageCategory.Tape);
@@ -415,7 +415,7 @@ namespace BizHawk.Emulation.Cores.Computers.SinclairSpectrum
 			}
 			// work out overall position within the tape
 			p = 0;
-			p = ((double)ourPos / (double)cnt) * 100.0;
+			p = (ourPos / (double)cnt) * 100.0;
 			sb.Append(p.ToString("N0", NumberFormatInfo.InvariantInfo) + "%");
 			SendMessage(sb.ToString().TrimEnd('\n'), MessageCategory.Tape);
 		}
